Feat: Create edit product function
