+++ admin/js/admin.js
@@ -3,7 +3,7 @@ import { renderAlltransacs } from './renderAlltransacs.js';
 import { deleteItem } from './delete.js';
 import { renderAdminSummary } from './admin-summary.js';
 import { addItem } from './addProduct.js';
-import { router } from './adminRouter.js';
+// import { router } from './adminRouter.js';
 
 (async () => {
   await renderAllProduct();
